+++ src/worker.js
@@ -1,3 +1,5 @@
+import {httpServerHandler} from 'cloudflare:node';
 import app from '../app.js';
 
-export default app;
+app.listen(3000);
+export default httpServerHandler({port: 3000});